เพิ่มระบบ viewdraw ได้โดยไม่ต้องผ่านหน้า write
เพิ่ม href เมื่อไม่ได้ login ใน drawinsert.php
เพิ่มไฟล์ viewdraw.php คือระบบสำหรับดูภาพประกอบได้โดยไม่ต้อง login

+++ public_html/drawinsert.php
@@ -2,6 +2,13 @@
 <? 
 //<!--  เพิ่มตาราง usedraw เข้ามาใหม่ ใน tbluser  -->
 session_start();
+echo '<title> write your story | Jaipun</title>';
+if($_SESSION["user_id"] == "")
+{
+  echo"<script>window.location.href='/register.php';</script>";
+  //header("location:register.php");
+  exit();
+}
 require('db_config.php');
 $conn->query("DELETE FROM `tbldrawuse` WHERE `tbldrawuse`.`user_id` = $_SESSION[user_id]");
   $drawname_id = $_GET[drawname_id];
